Continue refactoring of non-Qt-specific Matplotlib viewer code
